+++ src/egon/data/datasets/scenario_parameters/__init__.py
@@ -207,7 +207,6 @@ def insert_scenarios():
 
     session.commit()
 
-
 def get_sector_parameters(sector, scenario=None):
     """Returns parameters for each sector as dictionary.
 
@@ -296,7 +295,7 @@ class ScenarioParameters(Dataset):
     This dataset creates and fills a table in the database that includes central parameters
     for each scenarios. These parameters are mostly from extrernal sources, they are defined
     and referenced within this dataset.
-    The table is acced by various datasets to access the parameters for all sectors.
+    The table is accessed by various datasets to access the parameters for all sectors.
 
 
     *Dependencies*
@@ -312,7 +311,7 @@ class ScenarioParameters(Dataset):
     #:
     name: str = "ScenarioParameters"
     #:
-    version: str = "0.0.24"
+    version: str = "0.0.25"
 
     sources = DatasetSources(
         urls={